pyproject.toml
[project]
name = "websearch"
version = "0.1.0"
description = "simple web search worker"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.8",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.3",
    "python-dotenv>=1.0.1",
]
